Change the source of the queryid field
As of version 16, use the query_id field coming from
pg_stat_activity as input for the
pg_active_session_history.queryid one.

Indeed, with the changes done in 16 related to query
jumbling and pg_stat_statements, the computation done
in get_parsedinfo could differ from the one reported
in pg_stat_statements (so get rid of the one computed
in get_parsedinfo as of 16).